dir: move data use case to separate directory
